+++ app/commands/__init__.py
@@ -17,12 +17,3 @@ class CommandHandler:
             self.commands[command_name].execute(*args, **kwargs)
         except KeyError:
             print(f"No such command: {command_name}")
-
-class MenuCommand(Command):
-    def __init__(self, command_handler):
-        self.command_handler = command_handler
-
-    def execute(self):
-        print("Available Commands:")
-        for command_name in self.command_handler.commands:
-            print(f"- {command_name}")
